+++ src/main/java/com/ontarget/api/dao/DocumentGridKeyValueDAO.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.ontarget.bean.DocumentGridKeyValueDTO;
 
 public interface DocumentGridKeyValueDAO extends GenericDAO<DocumentGridKeyValueDTO>{
-	List<DocumentGridKeyValueDTO> getByDocumentId(long documentId);
-	List<DocumentGridKeyValueDTO> getByDocumentIdAndGridId(long documentId, String gridId);
-	boolean updateValue(long documentId, String gridId, int gridRowIndex, String key, String newValue, int modifiedBy);
+	List<DocumentGridKeyValueDTO> getByDocumentId(int documentId);
+	List<DocumentGridKeyValueDTO> getByDocumentIdAndGridId(int documentId, String gridId);
+	boolean updateValue(int documentId, String gridId, int gridRowIndex, String key, String newValue, int modifiedBy);
 }
