+++ src/main/java/com/ontarget/api/dao/DocumentKeyValueDAO.java
@@ -6,6 +6,6 @@ import java.util.List;
 import com.ontarget.bean.DocumentKeyValueDTO;
 
 public interface DocumentKeyValueDAO extends GenericDAO<DocumentKeyValueDTO>{
-	List<DocumentKeyValueDTO> getByDocumentId(long documentId);
-	boolean updateValue(long documentId, String key, String newValue, int modifiedBy);
+	List<DocumentKeyValueDTO> getByDocumentId(int documentId);
+	boolean updateValue(int documentId, String key, String newValue, int modifiedBy);
 }
